# backend/swagger/api-docs.yml
openapi: 3.0.0
info:
  title: ZeYang API
  description: ZeYang 房地產專案管理系統 API 文檔
  version: 2.1.0
  contact:
    name: API Support
    email: dev@ZeYang.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/api
    description: Development server
  - url: https://api.ZeYang.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            details:
              type: string
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          description: 標籤 ID
        identifier:
          type: string
          description: 標籤識別碼
        name:
          type: string
          description: 標籤名稱
        nameEn:
          type: string
          description: 英文名稱
        category:
          type: string
          enum: [style, feature, location, type, other]
          description: 標籤分類
        description:
          type: string
          description: 描述
        color:
          type: string
          description: 顏色代碼
        icon:
          type: string
          description: 圖標
        usageCount:
          type: integer
          description: 使用次數
...
paths:
...
  /tags:
    get:
      tags:
        - Tags
      summary: 取得標籤列表
      parameters:
        - in: query
          name: orderBy
          schema:
            type: string
            enum: [usageCount, name]
            default: usageCount
        - in: query
          name: orderDir
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
            enum: [style, feature, location, type, other]
      responses:
        200:
          description: 標籤列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tag'
    
    post:
      tags:
        - Tags
      summary: 建立新標籤
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                description:
                  type: string
                  maxLength: 200
      responses:
        201:
          description: 標籤建立成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /tags/popular:
    get:
      tags:
        - Tags
      summary: 取得熱門標籤
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        200:
          description: 熱門標籤列表
  /tags/search:
    get:
      tags:
        - Tags
      summary: 搜尋標籤
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        200:
          description: 搜尋結果
        400:
          description: 搜尋關鍵字不能為空
  /tags/{identifier}:
    get:
      tags:
        - Tags
      summary: 取得單一標籤
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: 標籤詳細資訊
        404:
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - Tags
      summary: 更新標籤
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                description:
                  type: string
                  maxLength: 200
      responses:
        200:
          description: 標籤更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - Tags
      summary: 刪除標籤
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: 標籤刪除成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'

  # 聯絡我們端點
...
  /settings:
    get:
      tags:
        - Settings
      summary: 取得所有設定或特定分類設定
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: 設定分類
      responses:
        200:
          description: 設定列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
    
    put:
      tags:
        - Settings
      summary: 批量更新設定
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - settings
              properties:
                settings:
                  type: object
                  description: 設定物件
      responses:
        200:
          description: 設定更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /settings/{key}:
    get:
      tags:
        - Settings
      summary: 取得單一設定
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
          description: 設定鍵值
      responses:
        200:
          description: 設定詳情
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
  /settings/smtp/test:
    post:
      tags:
        - Settings
      summary: 測試 SMTP 連線
      security:
        - bearerAuth: []
      responses:
        200:
          description: SMTP 測試成功
        400:
          description: SMTP 測試失敗
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /settings/category/email:
    get:
      tags:
        - Settings
      summary: 取得電子郵件設定
      security:
        - bearerAuth: []
      responses:
        200:
          description: 電子郵件設定
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
    
    put:
      tags:
        - Settings
      summary: 更新電子郵件設定
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                smtp_enabled:
                  type: boolean
                smtp_host:
                  type: string
                smtp_port:
                  type: integer
                  minimum: 1
                  maximum: 65535
                smtp_secure:
                  type: boolean
                smtp_username:
                  type: string
                smtp_password:
                  type: string
                smtp_from_email:
                  type: string
                  format: email
                smtp_from_name:
                  type: string
                admin_notification_emails:
                  type: string
                send_admin_notifications:
                  type: boolean
                send_user_confirmations:
                  type: boolean
      responses:
        200:
          description: 電子郵件設定更新成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/ValidationError'
    
    delete:
      tags:
        - Settings
      summary: 刪除電子郵件設定
      security:
        - bearerAuth: []
      responses:
        200:
          description: 電子郵件設定刪除成功
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'

  # 專案圖片管理端點
...
components:
  responses:
    Unauthorized:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 權限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 資源未找到
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 驗證錯誤
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
...